Menu created & function call added to main.cpp

+++ MileageCalculator.h
@@ -3,4 +3,4 @@
 
 using namespace std;
 
-double MileageCalculator(int, int, double);
+double MileageCalculator(double, double, double);
